refactor: postDetailStore를 postStore로 이름을 수정함
