+++ study-app/src/components/comment/CommentCreator.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useState, useContext } from 'react'
 
-import { PostContext, PostContextProps } from '../../stores/PostDetailStore'
+import { PostContext, PostContextProps } from '../../stores/PostStore'
 import { UserContext } from '../../stores/UserStore'
 
 import { CommentInputBox } from './CommentInputBox'
