+++ study-app/src/pages/PostUpdate.tsx
@@ -9,7 +9,7 @@ import { Input } from '../style-components/input/Input'
 import { Redirect, useParams } from 'react-router-dom'
 import { useFetch } from '../custom-hook/useFetch'
 import { PostDetail } from '../models/PostDetail'
-import { PostContextProps, PostContext } from '../stores/PostDetailStore'
+import { PostContextProps, PostContext } from '../stores/PostStore'
 
 const HeaderWrapper = styled.div`
   .selectBox__wrapper {
